add springdoc to backend and move java files to different folder
