+++ backend/src/main/java/com/hufudb/onedb/backend/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.hufudb.onedb.backend;
+package com.hufudb.onedb.backend.controller;
 
 import java.util.List;
 import java.util.Set;
